Use size_t in DEFINE_DATAINDEX macro and finalize features
- Update DEFINE_DATAINDEX macro in PropertyMarcos.h to take index of type size_t.
- Add DEFINE_STRUCTSIZE macro inside PropertyMarcos.h.
- Fix const-correctness compile error in DEFINE_MEMBER macro in PropertyMarcos.h.
- Update FUObjectItem definition in UObjectArray.h using DEFINE_MEMBER.
- Overwrite Array.h with complete TArray class featuring helper methods (GetData, Num, Max, operator[], etc.).
- Implement class FString inheriting from TArray<wchar_t> in UnrealString.h.
- Implement FString::ToString() featuring Epic Games / standard-SDK style wide-to-multibyte string constructor for UE4.12 x86.
- Add FMemory static memory helper class.
- Add VersionInfo class with static double EngineVersion and static int CL fields and InitParseVersion() to dynamically parse UE4/UE5 version strings.
- Update Logger::Init to check and clear existing log files.
- Update Address.h, pch.h, and VCXPROJ files cleanly.

+++ FortGen/Engine/Source/Runtime/Core/Public/Containers/PropertyMarcos.h
@@ -9,4 +9,4 @@
     static int32_t GetStructSize() { return Size; }
 
 #define DEFINE_DATAINDEX(Name, Offset, Type, StructSize) \
-    Type Get##Name(int32_t Index) const { return reinterpret_cast<Type>(*reinterpret_cast<uintptr_t*>(reinterpret_cast<uintptr_t>(this) + Offset) + StructSize * Index); }
+    Type Get##Name(size_t Index) const { return reinterpret_cast<Type>(*reinterpret_cast<uintptr_t*>(reinterpret_cast<uintptr_t>(this) + Offset) + StructSize * Index); }
